requirements.txt
django==3.2.2
djangorestframework==3.12.4
psycopg2==2.8.6
paho-mqtt
django-cors-headers